fix: vote message proto conversion and re-organise protos/conversions
