+++ applications/tari_validator_node/proto/dan/network.proto
@@ -24,6 +24,17 @@ syntax = "proto3";
 package tari.dan.network;
 
 import "common.proto";
+import "consensus.proto";
+import "transaction.proto";
+
+message DanMessage {
+  oneof message {
+    tari.dan.consensus.HotStuffMessage hot_stuff = 1;
+    tari.dan.consensus.VoteMessage vote = 2;
+    tari.dan.transaction.Transaction new_transaction = 3;
+    NetworkAnnounce network_announce = 4;
+  }
+}
 
 message NetworkAnnounce {
   bytes identity = 1;
@@ -35,32 +46,4 @@ message IdentitySignature {
   uint32 version = 1;
   tari.dan.common.Signature signature = 2;
   int64 updated_at = 3;
-}
-
-message GetPeersRequest {
-  uint64 since = 1;
-}
-
-message GetPeersResponse {
-  bytes identity = 1;
-  repeated bytes addresses = 2;
-  IdentitySignature identity_signature = 3;
-}
-
-message VNStateSyncRequest {
-  // Specific shard ids
-  repeated tari.dan.common.ShardId missing_shard_state_ids = 1;
-}
-
-message VNStateSyncResponse {
-  tari.dan.common.SubstateState substate_state = 1;
-}
-
-message GetVnStateInventoryRequest {
-  tari.dan.common.ShardId start_shard_id = 1;
-  tari.dan.common.ShardId end_shard_id = 2;
-}
-
-message GetVnStateInventoryResponse {
-  repeated tari.dan.common.ShardId inventory = 1;
 }
