+++ applications/tari_validator_node/src/p2p/proto/conversions/mod.rs
@@ -1,15 +1,15 @@
-//  Copyright 2022, The Tari Project
+//  Copyright 2024, The Tari Project
 //
 //  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
 //  following conditions are met:
 //
-//  1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
+//  2. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
 //  disclaimer.
 //
-//  2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
+//  3. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
 //  following disclaimer in the documentation and/or other materials provided with the distribution.
 //
-//  3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
+//  4. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
 //  products derived from this software without specific prior written permission.
 //
 //  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
@@ -19,25 +19,8 @@
 //  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-mod conversions;
 
-pub mod transaction {
-    include!(concat!(env!("OUT_DIR"), "/tari.dan.transaction.rs"));
-}
-
-pub mod consensus {
-    include!(concat!(env!("OUT_DIR"), "/tari.dan.consensus.rs"));
-}
-
-pub mod network {
-    #![allow(clippy::large_enum_variant)]
-    include!(concat!(env!("OUT_DIR"), "/tari.dan.network.rs"));
-}
-
-pub mod rpc {
-    include!(concat!(env!("OUT_DIR"), "/tari.dan.rpc.rs"));
-}
-
-pub mod common {
-    include!(concat!(env!("OUT_DIR"), "/tari.dan.common.rs"));
-}
+mod common;
+mod consensus;
+mod network;
+mod transaction;
