+++ applications/tari_validator_node/src/p2p/services/rpc_client.rs
@@ -37,7 +37,7 @@ use tari_dan_engine::transaction::Transaction;
 use tokio_stream::StreamExt;
 
 use crate::p2p::{
-    proto::{network::GetPeersRequest, validator_node::SubmitTransactionRequest},
+    proto::rpc::{GetPeersRequest, SubmitTransactionRequest},
     rpc,
 };
